Fix clang warnings for android builds (#6141)
https://github.com/youtube/cobalt/pull/6139 will actually remove the
clang warning suppression flags after this is submitted.

Bug: 374302273
Bug: 378934120

original-hexsha: 4f4cd5752ad884352d2ed528dea8e1f1d05ec380

+++ starboard/nplb/posix_compliance/posix_usleep_test.cc
@@ -113,7 +113,7 @@ TEST_F(PosixUsleepTests, SuccessfulSleep) {
                     << ", errno: " << errno << " (" << strerror(errno) << ")";
 
   long elapsed_us = TimevalDiffToMicroseconds(&start_time, &end_time);
-  EXPECT_GE(elapsed_us, kTestSleepUs)
+  EXPECT_GE(elapsed_us, static_cast<long>(kTestSleepUs))
       << "Sleep duration was too short. Requested: " << kTestSleepUs
       << "us, Elapsed: " << elapsed_us << "us.";
 }
