+++ starboard/shared/starboard/accessibility_internal.h
@@ -47,7 +47,7 @@ SbAccessibilityCaptionColor GetClosestCaptionColorRGB(int r, int g, int b) {
   int min_distance = std::numeric_limits<int>::max();
 
   // Find the reference color with the least distance (squared).
-  for (int i = 0; i < SB_ARRAY_SIZE(kRgbColors); i++) {
+  for (size_t i = 0; i < SB_ARRAY_SIZE(kRgbColors); i++) {
     int r_ref = 0xFF & (kRgbColors[i] >> 16);
     int g_ref = 0xFF & (kRgbColors[i] >> 8);
     int b_ref = 0xFF & (kRgbColors[i]);
@@ -95,7 +95,7 @@ int FindClosestReferenceValue(int value,
   int result = reference[0];
   int min_difference = std::numeric_limits<int>::max();
 
-  for (int i = 0; i < reference_size; i++) {
+  for (size_t i = 0; i < reference_size; i++) {
     int difference = abs(reference[i] - value);
     if (difference < min_difference) {
       result = reference[i];
